feat: Cleaned up code
